osxaudio: Support dynamic device switching
osxaudiosink will always try to switch on the fly, without
reinitializing anything. Usually works, AudioUnit will happily take care
of resampling and conversion in this case. If for some reason the switch
fails, we attempt to go back to the old device, and only error out if
that's no longer possible.

osxaudiosrc behaves differently because AudioUnitRender() used in our
ioproc doesn't want to resample anything. Due to that we have to set a
flag and renegotiate in create(), recreating the ringbuffer if needed.
Of course if device caps match then we don't reinitialize anything and
it behaves similarly to the sink variant.